requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
ansible==7.6.0
ansible-core==2.14.6
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Automat==20.2.0
autopep8==2.0.1
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.2.0
cmake==3.27.0
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
dbus-python==1.2.16
decorator==4.4.2
deep-translator==1.11.4
distlib==0.3.4
distro==1.5.0
docutils==0.17.1
Emailscrapy @ git+https://github.com/robertzengcn/emailscrapy.git@501632b3f3825cc1d65dfec9ddedcba6c08d622c
executing==0.8.3
fake-useragent==0.1.11
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.0
future==0.18.3
git-filter-repo==2.34.0
greenlet==1.1.2
h11==0.13.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imagesize==1.3.0
importlib-metadata==4.11.3
incremental==21.3.0
ipython==8.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
libvirt-python==7.0.0
lit==16.0.6
llvmlite==0.40.1
lxml==4.8.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.3
more-itertools==10.0.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.2
netaddr==0.8.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
outcome==1.1.0
packaging==21.3
parsel==1.6.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.5.1
proglog==0.1.10
prompt-toolkit==3.0.29
Protego==0.2.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
PyGObject==3.38.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0.1
queuelib==1.6.2
recommonmark==0.7.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
resolvelib==0.8.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
Scrapy==2.6.1
selenium==4.1.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.32
stack-data==0.2.0
sympy==1.12
terminator==2.1.0
tiktoken==0.3.1
tldextract==3.3.0
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
triton==2.0.0
Twisted==22.4.0
typing-extensions==4.7.1
urllib3==2.0.4
virtualenv==20.14.0
w3lib==1.22.0
wcwidth==0.2.5
wsproto==1.1.0
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
